backend/requirements.txt
daphne==4.1.2
Django==5.0.6
pillow==10.4.0
channels==4.1.0
requests==2.32.3
channels-redis==4.2.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1